refactor: tanstack query관련 로직 분리
