+++ src/queries/infiniteQueries.jsx
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from "@tanstack/react-query";
-import { getFeedsByPageNum } from "../api/feedApi";
+import { getFeedPages, getFeedsByPageNum } from "../api/feedApi";
 import useUserStore from "../store/useUserStore";
 import { queryKeys } from "./query.keys";
 
@@ -9,6 +9,17 @@ export const useMyFeedsInfiniteQuery = () => {
     queryKey: queryKeys.boardController.myFeeds(),
     queryFn: ({ pageParam = 1 }) => getFeedsByPageNum({ pageParam, userId: user.id }),
 
+    getNextPageParam: (lastPage, pages) => {
+      return lastPage.length === 5 ? pages.length + 1 : undefined;
+    },
+    select: (data) => data.pages.flat()
+  });
+};
+
+export const useAllFeedsInfiniteQuery = () => {
+  return useInfiniteQuery({
+    queryKey: queryKeys.boardController.feeds(),
+    queryFn: getFeedPages,
     getNextPageParam: (lastPage, pages) => {
       return lastPage.length === 5 ? pages.length + 1 : undefined;
     },
